requirements.txt
PyYAML
ddt
Appium-Python-Client==0.46
selenium==3.141.0
urllib3==1.25.3
